Added Security to the backend

+++ ppmtool/src/main/java/com/tool/ppmtool/service/impl/ProjectServiceImpl.java
@@ -14,9 +14,11 @@ import org.springframework.stereotype.Service;
 
 import com.tool.ppmtool.domain.Backlog;
 import com.tool.ppmtool.domain.Project;
+import com.tool.ppmtool.domain.User;
 import com.tool.ppmtool.exception.project.ProjectIdException;
 import com.tool.ppmtool.model.response.ProjectResponse;
 import com.tool.ppmtool.repository.ProjectRepository;
+import com.tool.ppmtool.repository.UserRepository;
 import com.tool.ppmtool.service.ProjectService;
 import com.tool.ppmtool.shared.dto.ProjectDTO;
 
@@ -25,23 +27,29 @@ public class ProjectServiceImpl implements ProjectService {
 
 	@Autowired
 	private ProjectRepository projectRepository;
+	
+	@Autowired
+	private UserRepository userRepository;
 
 
 	@Override
-	public ProjectDTO findProjectByIdentifier(String projectId) {
+	public ProjectDTO findProjectByIdentifier(String projectId, String username) {
 		Project project = projectRepository.findByProjectIdentifier(projectId.toUpperCase());
 		
 		if(project == null) 
 				throw new ProjectIdException("Project ID '" + projectId.toUpperCase() + "' does not exists");
 		
-		ProjectDTO returnValue = new ProjectDTO();
-		BeanUtils.copyProperties(project, returnValue);
-		return returnValue;
+		if(project.getProjectLeader().equals(username)) {
+			ProjectDTO returnValue = new ProjectDTO();
+			BeanUtils.copyProperties(project, returnValue);
+			return returnValue;
+		}
+		throw new ProjectIdException("Project ID '" + projectId.toUpperCase() + "' restricted");
 	}
 
 	@Override
-	public List<ProjectDTO> findAllProjects() {
-		Iterable<Project> storedProjects = projectRepository.findAll();
+	public List<ProjectDTO> findAllProjects(String principal) {
+		Iterable<Project> storedProjects = projectRepository.findAllByProjectLeader(principal);
 		List<Project> list = 
 				  StreamSupport.stream(storedProjects.spliterator(), false)
 				    .collect(Collectors.toList());
@@ -53,24 +61,45 @@ public class ProjectServiceImpl implements ProjectService {
 		
 		return returnValue;
 	}
+	
+//	 Shows ALL project in DB
+//	@Override
+//	public List<ProjectDTO> findAllProjects() {
+//		Iterable<Project> storedProjects = projectRepository.findAll();
+//		List<Project> list = 
+//				  StreamSupport.stream(storedProjects.spliterator(), false)
+//				    .collect(Collectors.toList());
+//		List<ProjectDTO> returnValue = new ArrayList<>();
+//		
+//		java.lang.reflect.Type listType = new TypeToken<List<ProjectDTO>>() {}.getType();
+//		ModelMapper modelMapper = new ModelMapper();
+//		returnValue = modelMapper.map(list, listType);
+//		
+//		return returnValue;
+//	}
 
 	@Override
-	public void deleteProjectByIdentifier(String projectId) {
+	public void deleteProjectByIdentifier(String projectId, String username) {
 		
 		Project project = projectRepository.findByProjectIdentifier(projectId.toUpperCase());
 		
 		if (project == null)
 			throw new ProjectIdException("No project with ID '" + projectId.toUpperCase() + "'");
 		
+		if(!project.getProjectLeader().equals(username)) 
+			throw new ProjectIdException("Project ID '" + projectId.toUpperCase() + "' restricted");
+		
 		projectRepository.deleteById(project.getId());
 	}
 
 
 	@Override
-	public ProjectDTO saveOrUpdateProject(ProjectDTO projectDTO) {
+	public ProjectDTO saveOrUpdateProject(ProjectDTO projectDTO, String username) {
 		Project project = new Project();
 		Project storedProjectDetails = new Project();
 		BeanUtils.copyProperties(projectDTO, project);
+		
+		
 		
 		project.setProjectIdentifier(project.getProjectIdentifier().toUpperCase());
 		storedProjectDetails = projectRepository.findByProjectIdentifier(project.getProjectIdentifier());
@@ -84,8 +113,13 @@ public class ProjectServiceImpl implements ProjectService {
 			backlog.setProjectIdentifier(project.getProjectIdentifier().toUpperCase());
 		}
 		
-		
+
+		//if user exists, add it to the project, user is logged
+		User user = userRepository.findByUsername(username);
+		project.setProjectLeader(user.getUsername());
+		project.setUser(user);
 		storedProjectDetails = projectRepository.save(project);
+
 		
 		ProjectDTO returnValue = new ProjectDTO();
 		BeanUtils.copyProperties(storedProjectDetails, returnValue);
@@ -94,7 +128,7 @@ public class ProjectServiceImpl implements ProjectService {
 	}
 
 	@Override
-	public ProjectDTO updateProject(ProjectDTO projectDTO) {
+	public ProjectDTO updateProject(ProjectDTO projectDTO, String username) {
 		Project newProject = new Project();
 		
 		BeanUtils.copyProperties(projectDTO, newProject);
@@ -104,6 +138,9 @@ public class ProjectServiceImpl implements ProjectService {
 		
 		if (projectEntity == null)
 			throw new ProjectIdException("No project with ID '" + newProject.getProjectIdentifier() + "'");
+		
+		if (!projectEntity.getProjectLeader().equals(username))
+			throw new ProjectIdException("Project ID '" + projectEntity.getProjectIdentifier().toUpperCase() + "' restricted");
 		
 		projectEntity.setDescription(newProject.getDescription());
 		projectEntity.setProjectName(newProject.getProjectName());
